Create FormatterFactory
Create a factory to call all formatters.

Create a constantize method to handle the instantiation of the class.
Since the `Object.const_get` brokes on ruby 1.9.3.

The constantize is the same of the ActiveSupport since we need only this
for the ActiveSupport, copied the method over add a new dependency.