Fix Travis mime-types dependency issue
Before this fix some of our Travis builds failed with:

    Gem::InstallError: mime-types-data requires Ruby version >= 2.0.

Idea for the fix was found here:
https://github.com/sgruhier/foundation_rails_helper/pull/116
Although we could not use the same method as I only want to lock
down the version of mime-type where necessary. Also trying to lock
the version to 2.6.2 would break our Rails-3.2 Appraisal path.