Update Contributing to reflect reality
Previously it seems that PRs were responded to within 1-3 days,
however this has not been the case for about a year so I'm
udating the relevant section of this guide in order to set more
realistic expectations. This is mainly so people do not get
discouraged when they see we haven't responded after a few days.