Make ActiveRecord specs easier to read
Extracting out the huge object strings so they do not hide the
actual test logic. The string are not important to see up front
as they are mostly in place for spotting regrssions. Having them
in seperate files also makes it easier to spot the differences
between the output of different parts of the dependency matrix.