Update changelog for the next release
I've also started adding links to the associated pull requests
and authors. The intention is to make it easier to find the
origin of a change.

The repetitive nature of the references annoys me, but
unfortunately I don't believe there is a cleaner way to do this
in markdown. If you can think of a solution, then please let us
know.

I've cleaned up some of the wording so the gramar is consistent
and sorted the changes from newest to oldest so that it's
consistent with the newest to oldest ordering of the version
numbers.

I've only made these changes for the latest commits, however
if somebody wishes to go back and add links for older changes,
that be a pull request I would happily merge.