Update urls to point to awesome-print organization
Before this commit a number or our urls were out of date which
lead to problems such as our build status on Travis being invalid
and people being asked to clone from an old version of the
repository.

Thanks to Michael Dvorkin for his many years of incredible work
on this project and for creating this organization so that the
project could live on with the help of others.