Add rubycop checking to codeclimate
According to the Contributing guide
(https://github.com/awesome-print/awesome_print/blob/master/CONTRIBUTING.md)
we encourage sticking to the Ruby style guide. However our code
climate setup does not use the Rubocop engine to check against
this style guide.

This commit enables the Rubocop checking. I'm not 100% sure if this
affects the code climate score, but even if it is I think this is
for the best as currently we are just hiding the issues.