Add some codeclimate recommended config
Codeclimate automatically generates a configuration and recommends
applying it. I have applied some of the options they specify.

The only thing I removed was the restriction of rubocop checks
which we are currently failing (I'd prefer they fail) and
additional languages such as php, python and javascript which
had strangely been added to the config.