Make option name more meaningful
I just noticed a code review which suggested against the name
we are currently using and gave some good reasoning:

"Naming things "new" is always a problem (since time passes and
it's no longer new). Maybe colonize_symbol_keys? ("colonize"
sounds kind of gross though...)
https://git.io/vKIiX

The ruby19_syntax name was suggested later in the conversation.

As we have not yet released with the previous option name, it's
now or never for this change.