Split out dotfile loading
This is so that we can stub the loading of the dotfile completely
instead of just trying to stub via it's implimentation. This also
makes it easier to only stub the dotfile loading instead of stubbing
all file access which then causes problems with other things, such
as Pry, which actually need to read from the filesystem.